fix migration for user_employee_service

+++ database/migrations/mod/2024_04_18_124329_add_column_title_to_user_employee_service_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('user_employee_service', function (Blueprint $table) {
-            Schema::table('service', function (Blueprint $table) {
+            Schema::table('user_employee_service', function (Blueprint $table) {
                 if (!Schema::hasColumn('user_employee_service', 'title')) {
                     $table->string('title')->comment('название');
                 }
@@ -30,7 +30,10 @@ return new class extends Migration
     public function down()
     {
         Schema::table('user_employee_service', function (Blueprint $table) {
-            $table->dropColumn('title');
+            if (Schema::hasColumn('user_employee_service', 'title')) {
+                $table->dropColumn('title');
+            }
+
         });
     }
 };
